MGS-3533 Convert from Spring Boot to RESTEasy/Undertow.
Cache health status of SAA API. The number of healthcheck requests
made to the healthcheck endpoint increased with recent changes to
the load balancer configuration.  This ensures that we limit the
number of upstream requests to the SAA API, despite the increased
healthcheck frequency. This change was requested by the providers
of the SAA API.